requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
CurrencyConverter==0.17.22
Django==5.0.6
djangorestframework==3.15.1
google_currency==1.0.10
idna==3.7
isodate==0.6.1
lxml==5.2.2
pillow==10.3.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
py-currency-converter==1.2.0
pyaes==1.6.1
pyasn1==0.6.0
pytz==2024.1
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.0
Telethon==1.23.0
urllib3==2.2.1
zeep==4.2.1
